Changed the way of getting the path of the data files, now we check first if there's an user defined version of the file and on fallback try it with the global one.
Now the execution log is written on the www directory and
notfied in the mail or RSS when a test cannot be executed properly.

Fixed makefile in order to make install work with /usr/local as prefix.

Updated INSTALL and changed VERSION to 0.4.1.

Added a mail template for fail notifications.

Many other little changes.


git-svn-id: https://pubsvn.warp.es/anste/trunk@112 18403268-a619-4aed-af80-3a7c80afbb3a